requirements.txt
# --- API / FastAPI
fastapi==0.110.1
uvicorn[standard]==0.30.6
httpx==0.27.2

# --- Config
pydantic==2.9.2
email-validator>=2.0.0
pydantic-settings==2.5.2

# --- DB
SQLAlchemy>=2.0
PyMySQL

# --- Encrypters
python-jose[cryptography]
bcrypt>=4.0.1
PyJWT

# --- Migrations
alembic>=1.13

# --- Tests
pytest>=8.3
pytest-cov>=5.0
python-multipart